Contextualize the frontpage for paying/door access
If you are a paying member, dont show the "how to become paying member"
If you are a paying member, but not the door circle, show information on
how to get door access
If you are paying and have door access, show the door buttons.

Texts are kinda just from the issue and improvised.

Fixes #41