Started cleaning up authz/account/profile
There is a bit of a mess now of account and authz-stuff all over client
and serverside. This is a start of moving from account -> profile client
side, and of making authorization a bit more explicit on the server
side. Fewer hard-coded queries and confusing `active_member` functinos
is the goal.