Add a "Print box label" button to "My profile"
The idea is that each member box at bitraf should have a label with a QR
code for quickly checking payment status, and with contact information
if payment has lapsed, making easier to keep the member storage clean
and with some available space.

This code adds a button to your personal profile that lets you print
such a label, by sending a message via MQTT with the information needed
to print a label. Elias is working on the component that will listen for
those messages and do the actual printing and label creation.

The idea of using MQTT is to keep p2k16 low on dependencies, and to not
have it do things that isnt "core".