Add support for different circles for each door
This allows us to have a set of circles for each door, and without
membership in that circle, you wont be able to open that door.

This check does not apply to employees of companies, only for paying
members.

If you attempt to open multiple doors, none will open if you lack any of
the circles.

For now all doors have `door` only, this is intended used at BV9 to
limit access to certain areas.

https://github.com/bitraf/infrastructure/issues/149