More correct door authz handling
We simplify a bit by checking payment and employee status first.
Employee have access to all dors, non-paying members have access to no
doors, so we can get that out of the way first.

Once we have done that, we check better that the account is a member of
at least one of the circles on a door, if there is no overlap of
circles, the whole request is denied.