fix: Use better profile list
When listing all accounts for /admin/account, the current query is very
slow because it 1) loads all data for each profile, and 2) the loading
of all profiles is done in a separate query.

As an easy fix, use the non-full version of the account list. With this
change the "fee" column is dropped.

Fixes #130.