Return a better error message when missing circle
A few places we require a specific circle to perform operations, but the
error message you get when that happens is a very generic "read only"
message, leaving the user confused.

This will make the error message a better "You are not a member of
circle `N`", leaving the user with a path forward.