Upgrade deps to get it running on python 3.13
Some hacks on cgi and crypt, which have been removed in core.

Some checks performed, but needs more testing. Some things can be
serialized wrongly. Before serializing an entity would end up as the ID,
now it will end up in an error, but most of those places can be changed
to `.id` to get it similar to earlier.