orm: Fixing mapping issue
Fixing this warning (and real problem) by updating a couple of types to
use the current pattern for declaring relationships.

> web/src/p2k16/core/models.py:642: SAWarning: relationship
'CompanyEmployee.company' will copy column company.id to column
company_employee.company, which conflicts with relationship(s):
'Company.employees' (copies company.id to company_employee.company).
If this is not the intention, consider if these relationships should be
linked with back_populates, or if viewonly=True should be applied to one
or more if they are read-only. For the less common case that foreign key
constraints are partially overlapping, the orm.foreign() annotation can
be used to isolate the columns that should be written towards.
To silence this warning, add the parameter 'overlaps="employees"' to the
'CompanyEmployee.company' relationship. (Background on this warning at:
https://sqlalche.me/e/20/qzyx)